Revert "refactor: move wrap-root-element to gatsby folder"
This reverts commit 9340f9e67fc68b8ed0d34d593337b1e0b9c3a2ae.
