+++ wrap-root-element.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { AppWrapper } from '../src/components/AppWrapper'
+import { AppWrapper } from './src/components/AppWrapper'
 
 export const wrapRootElement = ({ element }) => (
   <React.StrictMode>
